pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vlm-adversarial-patch-finder"
version = "0.1.0"
description = "AI robot model security demonstration system showcasing VLM adversarial vulnerabilities"
authors = [
    {name = "VicOne Research Team", email = "research@vicone.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
]
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0", 
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "huggingface-hub>=0.19.0",
    "datasets>=2.15.0",
    "tokenizers>=0.15.0",
    "opencv-python>=4.8.0",
    "scikit-image>=0.22.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0", 
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/rize0528/robo_adversarial_learning_kit"
Repository = "https://github.com/rize0528/robo_adversarial_learning_kit"
Issues = "https://github.com/rize0528/robo_adversarial_learning_kit/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "slow: Slow tests that may take more time",
]